backend/requirements.txt
google-genai>=0.5.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
opencv-python>=4.9.0
Pillow>=10.3.0
numpy>=1.26.0
reportlab>=4.2.0
torch>=2.2.0
torchvision>=0.17.0
